+++ CHANGELOG.rst
@@ -6,9 +6,8 @@ v9.4.1
 
 Added
 ~~~~~
-- optional predecessor_id argument in commands.launch. If passed, will print to 
+- optional predecessor_id argument in commands.launch. If passed, will print to
 stdout the preceding run link.
-
 
 v9.4.0
 ------